Replace Victory container component
This commit replaces the VictoryVoronoiContainer with the VictoryZoomContainer.
We used the VictoryVoronoiContainer to easily render tooltips but that container
does not work well if there are lots of sites. The sites either render on top
of each other or the sites outside the view are cut off. The VictoryZoomContainer
allows horizontal scrolling to display the sites but we need another solution
for rendering tooltips (upcoming).