Fix "Cannot read property 'map` of undefined" bug
This commit fixes a bug in which the user saw the message
"Cannot read property 'map` of undefined" when viewing a chart. This bug
surfaces when the assessment does not exist, causing our `largestHorizontalSection`
to be undefined. To fix this we default that value to an empty array.

This commit also includes the handling of empty data on the front-end with
a link to edit the chart.